refactor CJS
