+++ src/js/builtins/BunBuiltinNames.h
@@ -154,7 +154,7 @@ using namespace JSC;
     macro(lazy) \
     macro(lazyStreamPrototypeMap) \
     macro(lineText) \
-    macro(loadCJS2ESM) \
+    macro(loadEsmIntoCjs) \
     macro(localStreams) \
     macro(main) \
     macro(makeAbortError) \
